Do not remove plugin on cleaning empty placeholder
This should fix #98

The embed plugin insert a div before the placeholder:

```
add : function ($placeholder) {
    var formHtml = '<div class="medium-editor-toolbar...
    $(formHtml).appendTo($placeholder.prev());
```

The fix for Firefox that cleanup empty placeholder (related to #28)
has to check if this div exist before removing it.
I've made it kind of generic by checking for a div element instead of a class.