Handle bad embed
On medium.com if we type a bad embed data, it's converted in a normal paragraph.
So I've handle it the same way.

If we fail to retrieve content for an embed, instead of showing an ugly message to the user,
we just convert the content into a normal paragraph